requirements.txt
Django==3.1.7
django-cors-headers==3.7.0
channels==3.0.3
channels-redis==3.2.0
mysqlclient==2.2.0
PyJWT==2.1.0
whitenoise==5.2.0
django-bootstrap-datepicker-plus==3.0.5
django-froala-editor==3.2.1
django-inlinecss==0.3.0
django-nested-admin==3.3.3
django-sortedm2m==3.0.2
daphne==3.0.2
django-filter==2.4.0
djangorestframework==3.12.4
python-dateutil==2.8.1
numpy==1.21.6
pandas==1.3.5
redis==3.5.3
gunicorn==20.1.0
Pillow==10.0.0
pymysql==1.1.0
drf-yasg==1.20.0
reportlab==4.0.4
xlsxwriter==3.1.9
